Don't assert known suffix for wildcard match
Where the domain was only successfully matched by the wildcard match
rule, we shouldn't assert that the domain has a known suffix, nor that
it is definitely an ICANN or Private domain.

This could also arguably be applied to the 'registrable' portion of the
Domain, given a wildcard only match tells us nothing about the
registrability of the last token. However since the registrable field is
conflated with the root of the domain, we leave it alone so fewer changes
are required elsewhere in the library.